+++ packages/email/src/index.ts
@@ -8,6 +8,8 @@ export { emailTheme, type EmailTheme } from './theme'
 
 // Primitives.
 export { EmailLayout, type EmailLayoutProps } from './primitives/EmailLayout'
+export { EmailHeader, type EmailHeaderProps } from './primitives/EmailHeader'
+export { BrandMark, type BrandMarkProps, type BrandMarkColor } from './primitives/BrandMark'
 export { EmailButton, type EmailButtonProps } from './primitives/EmailButton'
 export { EmailHeading, type EmailHeadingProps } from './primitives/EmailHeading'
 export { EmailText, type EmailTextProps } from './primitives/EmailText'
@@ -16,6 +18,7 @@ export {
   EmailFooter,
   type EmailFooterProps,
   type EmailFooterLink,
+  type EmailFooterLabels,
 } from './primitives/EmailFooter'
 
 // Ready-made templates.